% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_there.R
\name{hello_there}
\alias{hello_there}
\title{print welcome message}
\usage{
hello_there(name)
}
\arguments{
\item{name}{A character string}
}
\value{
A character string containing a greeting
}
\description{
print welcome message
}
\examples{
hello_there("Susan")
}
